--- src/aiida_shell/engine/__init__.py
@@ -1,4 +1,0 @@
-"""Module for :mod:`aiida_shell.engine`."""
-from .launchers import launch_shell_job
-
-__all__ = ('launch_shell_job',)
