--- src/aiida_shell/engine/launchers/__init__.py
@@ -1,4 +1,0 @@
-"""Module for :mod:`aiida_shell.engine.launchers`."""
-from .shell_job import launch_shell_job
-
-__all__ = ('launch_shell_job',)
